Remove some leftovers in transfers view
