+++ pynicotine/gtkgui/downloads.py
@@ -124,15 +124,6 @@ class Downloads(TransferList):
         frame.ExpandDownloads.connect("toggled", self.OnExpandDownloads)
         self.OnExpandDownloads(None)
 
-    def saveColumns(self):
-        columns = []
-        widths = []
-        for column in self.frame.DownloadList.get_columns():
-            columns.append(column.get_visible())
-            widths.append(column.get_width())
-        self.frame.np.config.sections["columns"]["download_columns"] = columns
-        self.frame.np.config.sections["columns"]["download_widths"] = widths
-
     def OnTryClearQueued(self, widget):
 
         direction = "down"
@@ -263,26 +254,6 @@ class Downloads(TransferList):
         self.select_transfers()
         self.OnAbortTransfer(widget, False)
 
-    def OnSelectUserTransfer(self, widget):
-
-        if len(self.selected_users) == 0:
-            return
-
-        selected_user = widget.get_parent().user
-
-        sel = self.frame.DownloadList.get_selection()
-        fmodel = self.frame.DownloadList.get_model()
-        sel.unselect_all()
-
-        for item in self.transfers:
-            user_file, iter, transfer = item
-            user, filepath = user_file
-            if selected_user == user:
-                ix = fmodel.get_path(iter)
-                sel.select_path(ix,)
-
-        self.select_transfers()
-
     def on_key_press_event(self, widget, event):
 
         key = Gdk.keyval_name(event.keyval)
@@ -334,67 +305,6 @@ class Downloads(TransferList):
 
             if playfile:
                 executeCommand(executable, playfile, background=False)
-
-    def OnPopupMenuUsers(self, widget):
-
-        self.select_transfers()
-
-        self.popup_menu_users.clear()
-
-        if len(self.selected_users) > 0:
-
-            items = []
-            self.selected_users.sort(key=str.lower)
-
-            for user in self.selected_users:
-
-                popup = PopupMenu(self.frame, False)
-                popup.setup(
-                    ("#" + _("Send _message"), popup.OnSendMessage),
-                    ("#" + _("Show IP a_ddress"), popup.OnShowIPaddress),
-                    ("#" + _("Get user i_nfo"), popup.OnGetUserInfo),
-                    ("#" + _("Brow_se files"), popup.OnBrowseUser),
-                    ("#" + _("Gi_ve privileges"), popup.OnGivePrivileges),
-                    ("", None),
-                    ("$" + _("_Add user to list"), popup.OnAddToList),
-                    ("$" + _("_Ban this user"), popup.OnBanUser),
-                    ("$" + _("_Ignore this user"), popup.OnIgnoreUser),
-                    ("#" + _("Select User's Transfers"), self.OnSelectUserTransfer)
-                )
-                popup.set_user(user)
-
-                items.append((1, user, popup, self.OnPopupMenuUser, popup))
-
-            self.popup_menu_users.setup(*items)
-
-        return True
-
-    def OnPopupMenuUser(self, widget, popup=None):
-
-        if popup is None:
-            return
-
-        menu = popup
-        user = menu.user
-        items = menu.get_children()
-
-        act = False
-        if len(self.selected_users) >= 1:
-            act = True
-
-        items[0].set_sensitive(act)
-        items[1].set_sensitive(act)
-        items[2].set_sensitive(act)
-        items[3].set_sensitive(act)
-
-        items[6].set_active(user in [i[0] for i in self.frame.np.config.sections["server"]["userlist"]])
-        items[7].set_active(user in self.frame.np.config.sections["server"]["banlist"])
-        items[8].set_active(user in self.frame.np.config.sections["server"]["ignorelist"])
-
-        for i in range(4, 9):
-            items[i].set_sensitive(act)
-
-        return True
 
     def DoubleClick(self, event):
